requirement.txt
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
Django==5.2
djangorestframework==3.16.0
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
sqlparse==0.4.4
Werkzeug==3.0.3
